+++ src/index.ts
@@ -1,3 +1,4 @@
+// Components
 export { default as fhiTheme } from './components/FhiTheme';
 export { default as FhiButton } from './components/FhiButton';
 export { default as FhiDetails } from './components/FhiDetails';
@@ -6,3 +7,7 @@ export { default as FhiHeader } from './components/FhiHeader';
 export { default as FhiTabBar } from './components/FhiTabBar';
 export { default as FhiAccordion } from './components/FhiAccordion';
 export { default as FhiToastProvider, useToast as useFhiToast } from './components/FhiToast';
+
+// Utilities
+export { default as apiClient } from './utils/apiClient';
+export { default as useApi } from './utils/useApi';